backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
motor==3.6.0
nanoid==2.0.0
emergentintegrations
